requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioop-lts==0.2.1
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
elevenlabs==1.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
jiter==0.9.0
openai==1.75.0
PyAudio==0.2.14
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
python-dotenv==1.1.0
requests==2.32.3
setuptools==79.0.0
sniffio==1.3.1
SpeechRecognition==3.14.2
standard-aifc==3.13.0
standard-chunk==3.13.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
websockets==15.0.1
